+++ lib/omniauth/strategies/bookingsync.rb
@@ -32,6 +32,12 @@ module OmniAuth
           end
         end
       end
+
+      # Fixes regression in omniauth-oauth2 v1.4.0 by
+      # https://github.com/intridea/omniauth-oauth2/commit/85fdbe117c2a4400d001a6368cc359d88f40abc7
+      def callback_url
+        options[:redirect_uri] || (full_host + script_name + callback_path)
+      end
     end
   end
 end